+++ app/src/main/java/com/example/parkour/data/model/uptdate/PerformanceUpdate.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PerformanceUpdate(
     val status: String,
-    @SerialName("total_time") val totalTime: Double
+    @SerialName("total_time") val totalTime: Int
 )
